+++ de.uks.beast.model/src/model/impl/ServerImpl.java
@@ -103,7 +103,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 * @ordered
 	 */
-	protected static final long		RAM_EDEFAULT		= 0L;
+	protected static final int		RAM_EDEFAULT		= 0;
 	
 	/**
 	 * The cached value of the '{@link #getRam() <em>Ram</em>}' attribute.
@@ -113,7 +113,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 * @ordered
 	 */
-	protected long					ram					= RAM_EDEFAULT;
+	protected int					ram					= RAM_EDEFAULT;
 	
 	/**
 	 * The default value of the '{@link #getDiskSpace() <em>Disk Space</em>}' attribute.
@@ -123,7 +123,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 * @ordered
 	 */
-	protected static final long		DISK_SPACE_EDEFAULT	= 0L;
+	protected static final int		DISK_SPACE_EDEFAULT	= 0;
 	
 	/**
 	 * The cached value of the '{@link #getDiskSpace() <em>Disk Space</em>}' attribute.
@@ -133,7 +133,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 * @ordered
 	 */
-	protected long					diskSpace			= DISK_SPACE_EDEFAULT;
+	protected int					diskSpace			= DISK_SPACE_EDEFAULT;
 	
 	/**
 	 * The default value of the '{@link #getName() <em>Name</em>}' attribute.
@@ -290,7 +290,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public long getRam()
+	public int getRam()
 	{
 		return ram;
 	}
@@ -302,22 +302,21 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setRam(long newRam)
-	{
-		long oldRam = ram;
+	public void setRam(int newRam) {
+		int oldRam = ram;
 		ram = newRam;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, ModelPackage.SERVER__RAM, oldRam, ram));
 	}
-	
-	
-	
+
+
+
 	/**
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public long getDiskSpace()
+	public int getDiskSpace()
 	{
 		return diskSpace;
 	}
@@ -329,16 +328,15 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setDiskSpace(long newDiskSpace)
-	{
-		long oldDiskSpace = diskSpace;
+	public void setDiskSpace(int newDiskSpace) {
+		int oldDiskSpace = diskSpace;
 		diskSpace = newDiskSpace;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, ModelPackage.SERVER__DISK_SPACE, oldDiskSpace, diskSpace));
 	}
-	
-	
-	
+
+
+
 	/**
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
@@ -372,12 +370,10 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 */
 	public Network getNetwork() {
-		if (network != null && network.eIsProxy())
-		{
+		if (network != null && network.eIsProxy()) {
 			InternalEObject oldNetwork = (InternalEObject)network;
 			network = (Network)eResolveProxy(oldNetwork);
-			if (network != oldNetwork)
-			{
+			if (network != oldNetwork) {
 				if (eNotificationRequired())
 					eNotify(new ENotificationImpl(this, Notification.RESOLVE, ModelPackage.SERVER__NETWORK, oldNetwork, network));
 			}
@@ -406,8 +402,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	public NotificationChain basicSetNetwork(Network newNetwork, NotificationChain msgs) {
 		Network oldNetwork = network;
 		network = newNetwork;
-		if (eNotificationRequired())
-		{
+		if (eNotificationRequired()) {
 			ENotificationImpl notification = new ENotificationImpl(this, Notification.SET, ModelPackage.SERVER__NETWORK, oldNetwork, newNetwork);
 			if (msgs == null) msgs = notification; else msgs.add(notification);
 		}
@@ -422,8 +417,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 * @generated
 	 */
 	public void setNetwork(Network newNetwork) {
-		if (newNetwork != network)
-		{
+		if (newNetwork != network) {
 			NotificationChain msgs = null;
 			if (network != null)
 				msgs = ((InternalEObject)network).eInverseRemove(this, ModelPackage.NETWORK__SERVER, Network.class, msgs);
@@ -459,8 +453,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	{
 		Service oldService = service;
 		service = newService;
-		if (eNotificationRequired())
-		{
+		if (eNotificationRequired()) {
 			ENotificationImpl notification = new ENotificationImpl(this, Notification.SET, ModelPackage.SERVER__SERVICE, oldService, newService);
 			if (msgs == null) msgs = notification; else msgs.add(notification);
 		}
@@ -476,8 +469,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 */
 	public void setService(Service newService)
 	{
-		if (newService != service)
-		{
+		if (newService != service) {
 			NotificationChain msgs = null;
 			if (service != null)
 				msgs = ((InternalEObject)service).eInverseRemove(this, EOPPOSITE_FEATURE_BASE - ModelPackage.SERVER__SERVICE, null, msgs);
@@ -499,8 +491,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 */
 	@Override
 	public NotificationChain eInverseAdd(InternalEObject otherEnd, int featureID, NotificationChain msgs) {
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__NETWORK:
 				if (network != null)
 					msgs = ((InternalEObject)network).eInverseRemove(this, ModelPackage.NETWORK__SERVER, Network.class, msgs);
@@ -518,8 +509,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	 */
 	@Override
 	public NotificationChain eInverseRemove(InternalEObject otherEnd, int featureID, NotificationChain msgs) {
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__NETWORK:
 				return basicSetNetwork(null, msgs);
 			case ModelPackage.SERVER__SERVICE:
@@ -538,8 +528,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	@Override
 	public Object eGet(int featureID, boolean resolve, boolean coreType)
 	{
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__IP:
 				return getIp();
 			case ModelPackage.SERVER__CPU_AMOUNT:
@@ -571,8 +560,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	@Override
 	public void eSet(int featureID, Object newValue)
 	{
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__IP:
 				setIp((String)newValue);
 				return;
@@ -583,10 +571,10 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 				setCpuType((String)newValue);
 				return;
 			case ModelPackage.SERVER__RAM:
-				setRam((Long)newValue);
+				setRam((Integer)newValue);
 				return;
 			case ModelPackage.SERVER__DISK_SPACE:
-				setDiskSpace((Long)newValue);
+				setDiskSpace((Integer)newValue);
 				return;
 			case ModelPackage.SERVER__NAME:
 				setName((String)newValue);
@@ -611,8 +599,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	@Override
 	public void eUnset(int featureID)
 	{
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__IP:
 				setIp(IP_EDEFAULT);
 				return;
@@ -651,8 +638,7 @@ public class ServerImpl extends MinimalEObjectImpl.Container implements Server
 	@Override
 	public boolean eIsSet(int featureID)
 	{
-		switch (featureID)
-		{
+		switch (featureID) {
 			case ModelPackage.SERVER__IP:
 				return IP_EDEFAULT == null ? ip != null : !IP_EDEFAULT.equals(ip);
 			case ModelPackage.SERVER__CPU_AMOUNT:
